LayoutTests/imported/w3c/web-platform-tests/html/semantics/document-metadata/the-meta-element/pragma-directives/attr-meta-http-equiv-refresh/allow-scripts-flag-changing-1-expected.txt

CONSOLE MESSAGE: line 2566: Error: assert_unreached: The iframe from which the meta came from must not refresh Reached unreachable code

FAIL Meta refresh is blocked by the allow-scripts sandbox flag at its creation time, not when refresh comes due Error: assert_unreached: The iframe from which the meta came from must not refresh Reached unreachable code
